pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcodegen"
version = "0.1.1"
description = "Генератор G-кода (плоскость, круглый/квадратный карманы) + валидатор, RU i18n"
readme = "README.md"
authors = [{name = "Dubovik Vadim", email = "dubovikv88@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "typer>=0.9",
  "rich>=13.0"
]

[project.scripts]
gcodegen = "gcodegen.cli:app"

[tool.setuptools.packages.find]
where = ["gcodegen"]
